Update core.js
I think this change from "TURM" to "Turm" will solve my problem i recently wrote you per Mail. Because in ACT it is standing with small capitals. Guess it's the same with "DAME" to "Dame" and "LÄUFER" to "Läufer". Hope we can give it a try.

Greeting Jan